+++ docs/news.rst
@@ -5,13 +5,13 @@ News
 .. module :: downward.experiments
 
 
-v3.0 (unreleased)
+v3.0 (2018-07-10)
 -----------------
 
 lab
 ^^^
-* Add :meth:`exp.add_parser() <lab.experiment.Experiment.add_parser>` (Silvan). See also :ref:`parsing`.
-* Add :meth:`exp.add_parse_again_step() <lab.experiment.Experiment.add_parse_again_step>` for running parsers again (Silvan).
+* Add :meth:`exp.add_parser() <lab.experiment.Experiment.add_parser>` method. See also :ref:`parsing` (Silvan).
+* Add :meth:`exp.add_parse_again_step() <lab.experiment.Experiment.add_parse_again_step>` method for running parsers again (Silvan).
 * Require that the ``build``, ``start_runs`` and ``fetch`` steps are added explicitly (see :class:`~lab.experiment.Experiment`).
 * Remove *required* argument from ``add_resource()``. All resources are now required.
 * Use stricter naming rules for commands and resources. See respective ``add_*`` methods for details.